requirements.txt
requests==2.18.4
PyYAML==3.12
Flask==1.0.2